+++ miden-crypto/src/lib.rs
@@ -8,6 +8,7 @@ extern crate std;
 
 pub mod dsa;
 pub mod ecdh;
+pub mod encryption;
 pub mod hash;
 pub mod merkle;
 pub mod rand;